refactor(crossEntity): decompose CrossEntityAutomationView into sub-components
Extracts four self-contained sections into src/components/crossEntity/:
- SummaryStatsGrid: six summary stat cards
- RiskWarningsSection: high/medium risk warning cards
- PipelineTracesPanel: pipeline traces sub-view with filter + accordion
- GlobalChainMapPanel: chain map table sub-view with type/operation filters

CrossEntityAutomationView becomes a thin orchestrator managing only the
empty-state guard, the sub-view TabList, and delegation to the panels.
